requirements.txt
pyinstaller
pandas
xlsxwriter
openpyxl
xlrd
python-docx
bs4
jinja2
